not working edit and create user

+++ src/main/java/ru/kata/spring/boot_security/demo/controller/AdminController.java
@@ -33,7 +33,7 @@ public class AdminController {
     }
 
     @GetMapping("/{id}")
-    public String showUserByID(@PathVariable("id") Integer id, ModelMap modelMap) {
+    public String showUserByID(@PathVariable("id") Long id, ModelMap modelMap) {
         modelMap.addAttribute("user", userService.findUserById(id));
         System.out.println(userService.findUserById(id));
         return "/admin/show";
@@ -58,20 +58,20 @@ public class AdminController {
     }
 
     @GetMapping("/{id}/edit")
-    public String editUser(ModelMap modelMap, @PathVariable("id") Integer id) {
+    public String editUser(ModelMap modelMap, @PathVariable("id") Long id) {
         modelMap.addAttribute("user", userService.findUserById(id));
         modelMap.addAttribute("roles", roleService.getRoles());
         return "admin/admin_page";
     }
 
     @PatchMapping("/{id}")
-    public String updateUser(@ModelAttribute("user") User user, @PathVariable("id") Integer id) {
+    public String updateUser(@ModelAttribute("user") User user, @PathVariable("id") Long id) {
         userService.updateUser(id, user);
         return "redirect:/admin";
     }
 
     @DeleteMapping("/{id}")
-    public String deleteUser(@PathVariable("id") Integer id) {
+    public String deleteUser(@PathVariable("id") Long id) {
         userService.deleteUserById(id);
         return "redirect:/admin";
     }
