+++ src/main/java/ru/kata/spring/boot_security/demo/service/RoleServiceImpl.java
@@ -22,7 +22,7 @@ public class RoleServiceImpl implements RoleService{
         return roleRepository.findAll();
     }
     @Override
-    public Role findRoleById(Integer id) {
+    public Role findRoleById(Long id) {
         return roleRepository.findById(id).orElse(null);
     }
     @Override
